CLN: misplaced TimedeltaIndex tests (#31755)
