+++ pandas/tests/indexes/timedeltas/test_constructors.py
@@ -10,6 +10,12 @@ from pandas.core.arrays import TimedeltaArray
 
 
 class TestTimedeltaIndex:
+    @pytest.mark.parametrize("unit", ["Y", "y", "M"])
+    def test_unit_m_y_raises(self, unit):
+        msg = "Units 'M' and 'Y' are no longer supported"
+        with pytest.raises(ValueError, match=msg):
+            TimedeltaIndex([1, 3, 7], unit)
+
     def test_int64_nocopy(self):
         # GH#23539 check that a copy isn't made when we pass int64 data
         #  and copy=False
