+++ pandas/tests/indexes/timedeltas/test_timedelta.py
@@ -284,17 +284,3 @@ class TestTimedeltaIndex(DatetimeLike):
 
         result = td.astype("timedelta64[s]")
         tm.assert_index_equal(result, expected)
-
-    @pytest.mark.parametrize("unit", ["Y", "y", "M"])
-    def test_unit_m_y_raises(self, unit):
-        msg = "Units 'M' and 'Y' are no longer supported"
-        with pytest.raises(ValueError, match=msg):
-            TimedeltaIndex([1, 3, 7], unit)
-
-
-class TestTimeSeries:
-    def test_series_box_timedelta(self):
-        rng = timedelta_range("1 day 1 s", periods=5, freq="h")
-        s = Series(rng)
-        assert isinstance(s[1], Timedelta)
-        assert isinstance(s.iat[2], Timedelta)
